chore: split toast context and provider to make docs auto-generation possible
